requirements.txt
pymavlink==2.4.8
dronekit
dronekit-sitl
pyqt5
matplotlib
pyaml
python-fcl
pyvista
pyvistaqt
meshio